hard
Using the idea of binary search. To find the kth number in 2 sorted arrays, find the k/2th number in the arrays, that is nums1[k/2-1] and nums2[k/2-1]. If nums1[k/2-1] is bigger than nums2[k/2-1], the first k/2 elements is all smaller than the kth number, they can be deleted. 
The same works for the situation that nums1[k/2-1] is smaller than nums2[k/2-1]. When the nums1[k/2-1] equals nums2[k/2-1], nums1[k/2-1] is just the kth number.